+++ src/const.py
@@ -11,9 +11,9 @@ MAP_HEIGHT = 43
 FONT_PATH = 'data/font/cp437_12x12.png'
 FONT_FLAG = tcod.FONT_TYPE_GREYSCALE | tcod.FONT_LAYOUT_ASCII_INROW
 
-FOV_ALGORITHM = tcod.FOV_BASIC
+FOV_ALGORITHM = tcod.FOV_SHADOW
 FOV_LIGHT_WALLS = True
-FOV_RADIUS = 80  # 10
+FOV_RADIUS = 10
 
 
 COLORS = {
